Day 7, Part 2: Consolidate logic into Hand
If we treat J as Jack and 1 as Joker, and our logic can handle the
presence of both in a given hand, then we can solve both parts by simply
replacing J with 1 in the input for the second part to simulate the
different interpretation of the input. This eliminated a ton of
duplication between Hand and JokerHand, and made main very concise.